# docs/openapi/admin-search-v1.yaml
openapi: 3.0.3
info:
  title: RideBook Admin Search and Export API
  version: 1.0.0
  description: >-
    Frozen contract for paginated admin search, advanced filters, multi-field
    sorting, and CSV export. Date-time bounds are inclusive and use ISO-8601.
servers:
  - url: /RideBook
security:
  - bearerAuth: []
...
paths:
...
  /admin/customers:
    get:
      tags: [Admin Customers]
      operationId: searchAdminCustomers
      summary: Search and filter customers
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/Size'
        - name: sort
          in: query
          description: >-
            Comma-separated field:direction terms. Allowed fields:
            customerName, phone, email, birthDate, createdAt.
          schema: { type: string, default: 'customerName:asc' }
          example: 'createdAt:desc,customerName:asc'
        - name: search
          in: query
          description: Case-insensitive search across name, phone, email, and address.
          schema: { type: string, maxLength: 200 }
        - $ref: '#/components/parameters/AccountStatus'
        - name: genders
          in: query
          description: Comma-separated exact gender values, case-insensitive.
          style: form
          explode: false
          schema:
            type: array
            items: { type: string }
        - $ref: '#/components/parameters/CreatedFrom'
        - $ref: '#/components/parameters/CreatedTo'
      responses:
        '200':
          description: Matching customers.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CustomerPageEnvelope' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/customers/export:
    get:
      tags: [Admin Customers]
      operationId: exportAdminCustomers
      summary: Export filtered customers as UTF-8 CSV
      description: Uses the same filters and sort syntax as search; pagination is ignored.
      parameters:
        - $ref: '#/components/parameters/CustomerSort'
        - $ref: '#/components/parameters/CustomerSearch'
        - $ref: '#/components/parameters/AccountStatus'
        - $ref: '#/components/parameters/Genders'
        - $ref: '#/components/parameters/CreatedFrom'
        - $ref: '#/components/parameters/CreatedTo'
      responses:
        '200': { $ref: '#/components/responses/CsvExport' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
    Size:
      name: size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
...
    AccountStatus:
      name: accountStatus
      in: query
      schema: { type: boolean }
...
    CustomerSort:
      name: sort
      in: query
      schema: { type: string, default: 'customerName:asc' }
    CustomerSearch:
      name: search
      in: query
      schema: { type: string, maxLength: 200 }
    Genders:
      name: genders
      in: query
      style: form
      explode: false
      schema:
        type: array
        items: { type: string }
    CreatedFrom:
      name: createdFrom
      in: query
      schema: { type: string, format: date-time }
    CreatedTo:
      name: createdTo
      in: query
      schema: { type: string, format: date-time }
  responses:
    CsvExport:
      description: UTF-8 CSV attachment containing all rows matching the filters.
      headers:
        Content-Disposition:
          schema: { type: string }
          example: attachment; filename="bookings-2026-09-01.csv"
      content:
        text/csv:
          schema: { type: string, format: binary }
    BadRequest:
      description: Invalid page, range, enum, or sort field.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Unauthorized:
      description: Missing or invalid access token.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
    Forbidden:
      description: Caller does not have the ADMIN role.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
  schemas:
...
    PageMetadata:
      type: object
      required: [size, number, totalElements, totalPages]
      properties:
        size: { type: integer, minimum: 0 }
        number: { type: integer, minimum: 0 }
        totalElements: { type: integer, format: int64, minimum: 0 }
        totalPages: { type: integer, minimum: 0 }
...
    CustomerPageEnvelope:
      allOf:
        - $ref: '#/components/schemas/SuccessEnvelope'
        - type: object
          properties:
            result:
              type: object
              required: [content, page]
              properties:
                content:
                  type: array
                  items: { $ref: '#/components/schemas/Customer' }
                page: { $ref: '#/components/schemas/PageMetadata' }
    SuccessEnvelope:
      type: object
      required: [status, message, result]
      properties:
        status: { type: integer, example: 200 }
        message: { type: string }
        result: { type: object }
    ErrorEnvelope:
      type: object
      required: [status, message]
      properties:
        status: { type: integer }
        message: { type: string }
    Account:
      type: object
      properties:
        accountId: { type: string, format: uuid }
        userName: { type: string }
        accountStatus: { type: boolean }
        createdAt: { type: string, format: date-time }
        role:
          type: object
          properties:
            roleName: { type: string, enum: [ADMIN, CUSTOMER, DRIVER] }
            description: { type: string }
...
    Customer:
      type: object
      properties:
        customerId: { type: string, format: uuid }
        customerName: { type: string }
        phone: { type: string }
        address: { type: string, nullable: true }
        email: { type: string, format: email, nullable: true }
        gender: { type: string, nullable: true }
        avatar: { type: string, nullable: true }
        birthDate: { type: string, format: date, nullable: true }
        account: { $ref: '#/components/schemas/Account' }